refactor: extract rarity functionality from util
