+++ contracts/solana/cryptopets/programs/cryptopets/src/instructions/create_starter_pet.rs
@@ -2,8 +2,8 @@ use anchor_lang::prelude::*;
 
 use crate::{
     errors::ErrorCode,
+    rarity::Rarity,
     state::{GlobalState, PetAccount, PlayerProfile},
-    util::Rarity,
 };
 
 pub fn handler(ctx: Context<CreateStarterPet>, name: String, dna: u64, rarity: u8) -> Result<()> {
@@ -12,7 +12,7 @@ pub fn handler(ctx: Context<CreateStarterPet>, name: String, dna: u64, rarity: u
         ErrorCode::NameTooLong
     );
 
-    let rarity = Rarity::try_from_u8(rarity).ok_or(ErrorCode::InvalidRarity)?;
+    let rarity: Rarity = rarity.try_into()?;
 
     let global_state = &mut ctx.accounts.global_state;
     let player_profile = &mut ctx.accounts.player_profile;
@@ -37,7 +37,7 @@ pub fn handler(ctx: Context<CreateStarterPet>, name: String, dna: u64, rarity: u
     pet.id = pet_id;
     pet.owner = ctx.accounts.owner.key();
     pet.dna = dna;
-    pet.rarity = rarity.as_u8();
+    pet.rarity = rarity.into();
     pet.level = 1;
     pet.ready_time = Clock::get()?.unix_timestamp;
     pet.win_count = 0;
